# docs/swagger/swagger.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/EqualNine/Lab01/1.0.0
info:
  version: "1.0.0"
  title: ServerING API
  description: The API for the ServerING project
paths:
...
  '/api/v1/countries':
    get:
      tags:
          - Country
      description: Returns All Countries

      responses:
        '200':
          description: All countries

          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Country'
                
        '204':
            description: No Content
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No Countries"
    
    post:
      security:
        - bearerAuth: []      

      tags:
        - Country
      description: Post Country

      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Authorization Bearer

      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Russia
                levelOfInterest:
                  type: integer
                  example: 100
                overallPlayers:
                  type: integer
                  example: 15000

      responses:
        '201':
          description: Saved Country

          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Country'
                
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "User must be logged in"
        
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorMessage:
                    type: string
                    example: "Not enough rights for this"
  '/api/v1/countries/{id}':
    get:
      tags:
          - Country
      description: Return Country by Id

      parameters:
          - in: path
            required: true
            name: id
            schema:
              type: integer
            description: Country Id

      responses:
        '200':
          description: Country by Id

          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Country'
        
        '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No country with id: '5'"

    put:
        security:
          - bearerAuth: []
          
        tags:
          - Country

        parameters:
          - in: header
            name: Authorization
            required: true
            schema:
              type: string
            description: Authorization Bearer 
        
          - in: path
            name: id
            required: true
            schema:
              type: integer
            description: Country Id
        
        requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    example: Russia
                  levelOfInterest:
                    type: integer
                    example: 100
                  overallPlayers:
                    type: integer
                    example: 15000

        responses:
          '201':
            description: Updated Country

            content: 
              application/json:
                schema:
                  $ref: '#/components/schemas/Country'
          
          '401':
            description: Unauthorized
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "User must be logged in"
          
          '403':
            description: Forbidden
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "Not enough rights for this"
          
          '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No country with id: '5'"

    patch:
        security:
          - bearerAuth: []
          
        tags:
          - Country

        parameters:
          - in: header
            name: Authorization
            required: true
            schema:
              type: string
            description: Authorization Bearer 
        
          - in: path
            name: id
            required: true
            schema:
              type: integer
            description: Country Id
        
        requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    example: USSR 

        responses:
          '201':
            description: Updated Country 
            content: 
              application/json:
                schema:
                  $ref: '#/components/schemas/Country'

          '401':
            description: Unauthorized
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "User must be logged in"
          
          '403':
            description: Forbidden
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "Not enough rights for this"
          
          '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No country with id: '5'"

    delete:
        security:
          - bearerAuth: []

        tags:
          - Country
        description: Delete Country by Id

        parameters:
          - in: header
            name: Authorization
            required: true
            schema:
              type: string
            description: Authorization Bearer 
    
          - in: path
            name: id
            required: true
            schema:
              type: integer
            description: Country Id
                
        responses:
          '200':
            description: Country removed
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: "Country deleted successfully!"
          
          '401':
            description: Unauthorized
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "User must be logged in"
          
          '403':
            description: Forbidden
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "Not enough rights for this"
          
          '404':
            description: Not Found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    errorMessage:
                      type: string
                      example: "No country with id: '5'"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # только для документации
  schemas:
...
    Country:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Russia
        levelOfInterest:
          type: integer
          example: 100
        overallPlayers:
          type: integer
          example: 15000
...